working

+++ src/main/java/ru/spbu/math/plok/model/client/Distribution.java
@@ -29,7 +29,7 @@ public abstract class Distribution {
 
 	public abstract long getRandomLong(long from, long to);
 	
-	public abstract int getRandomInt(long from, long to);
+	public abstract int getRandomInt(int from, int to);
