+++ src/main/java/ru/spbu/math/plok/model/client/ExponentialDistribution.java
@@ -12,7 +12,7 @@ public class ExponentialDistribution extends Distribution{
 	}
 
 	@Override
-	public int getRandomInt(long from, long to) {
+	public int getRandomInt(int from, int to) {
 		return 0;
 	}
